backend/pyproject.toml
[project]
name = "transformed-roofing-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.4",
    "psycopg[binary]>=3.2",
    "python-decouple>=3.8",
    "dj-database-url>=2.3",
    "gunicorn>=22.0",
    "whitenoise>=6.8",
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "model-bakery>=1.19",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py"]
